Refactor `FlowSemantic`/`FlowNode`/`FlowPath` out of the FullNameSemantics parser
